refactor: split folder theme folder between color and writing
