+++ packages/ui/src/components/molecules/EditableInput/EditableInput.tsx
@@ -19,7 +19,7 @@
 import React, {useEffect, useMemo, useState} from 'react';
 import {StyleSheet, TouchableOpacity, View} from 'react-native';
 import {getCommonStyles} from '../../../utils/commons-styles';
-import {useThemeColor} from '../../../theme/ThemeContext';
+import {useThemeColor} from '../../../theme';
 import {Icon, Input} from '../../atoms';
 
 interface EditableInputProps {
