+++ packages/ui/src/components/molecules/FromTo/FromTo.tsx
@@ -18,7 +18,7 @@
 
 import React from 'react';
 import {View, StyleSheet} from 'react-native';
-import {useThemeColor} from '../../../theme/ThemeContext';
+import {useThemeColor} from '../../../theme';
 import {Icon} from '../../atoms';
 
 interface FromToProps {
