+++ packages/ui/src/theme/writing/index.ts
@@ -16,30 +16,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, {useMemo} from 'react';
-import {View, StyleSheet} from 'react-native';
-import {useThemeColor} from '../../../theme';
-
-interface HorizontalRuleProps {
-  style?: any;
-}
-
-const HorizontalRule = ({style}: HorizontalRuleProps) => {
-  const Colors = useThemeColor();
-
-  const styles = useMemo(() => {
-    return getStyles(Colors);
-  }, [Colors]);
-
-  return <View style={[styles.line, style]} />;
-};
-
-const getStyles = Colors =>
-  StyleSheet.create({
-    line: {
-      borderBottomColor: Colors.secondaryColor.background,
-      borderBottomWidth: 1,
-    },
-  });
-
-export default HorizontalRule;
+export {WritingThemeProvider, useWritingStyle} from './WritingThemeContext';
+export {
+  writingDefaultTheme,
+  Writing,
+  WritingStyles,
+  TextStyle,
+  useWritingType,
+} from './writingTheme';
